move jwt classes, unify implementation for oauth and openid
