+++ src/main/java/it/smartcommunitylab/aac/jwt/DefaultJWTEncryptionAndDecryptionService.java
@@ -1,4 +1,4 @@
-package it.smartcommunitylab.aac.openid.service;
+package it.smartcommunitylab.aac.jwt;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
@@ -31,6 +31,9 @@ import com.nimbusds.jose.jwk.ECKey;
 import com.nimbusds.jose.jwk.JWK;
 import com.nimbusds.jose.jwk.OctetSequenceKey;
 import com.nimbusds.jose.jwk.RSAKey;
+
+import it.smartcommunitylab.aac.jose.JWKSetKeyStore;
+
 
 /**
  * @author wkim
