+++ src/main/java/it/smartcommunitylab/aac/jwt/JWTEncryptionAndDecryptionService.java
@@ -1,4 +1,4 @@
-package it.smartcommunitylab.aac.openid.service;
+package it.smartcommunitylab.aac.jwt;
 
 import java.util.Collection;
 import java.util.Map;
