+++ src/main/java/it/smartcommunitylab/aac/jwt/assertion/SelfAssertionValidator.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  ******************************************************************************/
 
-package it.smartcommunitylab.aac.openid.service;
+package it.smartcommunitylab.aac.jwt.assertion;
 
 /**
  * @author raman
@@ -33,6 +33,8 @@ import com.nimbusds.jwt.JWT;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 
+import it.smartcommunitylab.aac.jwt.JWTSigningAndValidationService;
+
 /**
  * Validates all assertions generated by this server
  *
@@ -45,7 +47,8 @@ public class SelfAssertionValidator {
 
 	@Autowired
 	private JWTSigningAndValidationService jwtService;
-	@Value("${openid.issuer}")
+	
+	@Value("${jwt.issuer}")
 	private String issuer;
 
 	public boolean isValid(JWT assertion) {
